Link Audio: move the session settings into a Link settings tab
The Link page had grown into two unrelated halves: device-wide settings
(Link on/off, transport sync, Link name, latency, sync to incoming) and
the per-slot Receive/Send lists. The settings half belongs in Settings,
and the slot lists are about to move into the Devices pane, so split them
apart now.

Extract the pieces both halves share into components/linkAudio/ first --
LinkAudioNameInput and scrollInputIntoView, SlotControls, the source and
sink rows, and the add-slot menu and modal -- plus the reorder swap into
lib/linkAudio. No behavior change beyond hiding the reorder arrows on a
list too short to reorder, where they were inert controls anyway.

The new Link tab is a hand-built Tabs.Panel rather than a
SettingsTabPanel: Link state lives in state.linkAudio and the transport
store, not in app settings or the runner config, so it has no
ConfigRecord to render through. It borrows settings.module.css's row
classes so it lines up with the generic tabs. Receive health sits outside
that row layout -- three stats and a button don't fit itemInputWrap's
250px control column, which clipped the button.

Sources and sinks are "Receives" and "Sends" in the UI now, matching
jack_transport_link's renamed port group. Display only: the model,
actions and OSC paths keep jtl's own vocabulary.

The Link page keeps the slot lists until the Devices pane absorbs them,
at which point the page and its route go away.
